requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
requests==2.31.0
python-multipart==0.0.6
transformers==4.35.2
torch==2.1.1
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
bcrypt==4.0.1
